requirements.txt
numpy==2.2.4
tensorboard==2.19.0
torch==2.6.0
torchvision==0.21.0
pillow==11.1.0
transformer==4.51.2
diffusers==0.33.1
tqdm==4.67.1
ftfy==6.3.1
regex==2024.11.6
clip==1.0

